+++ src/components/Nav/SearchBar/Modal.component.js
